panacea-fetch -> just-fetch, refactor folder structure a bit
