+++ packages/just-fetch/README.md
@@ -1,11 +1,11 @@
-# @foxglove/panacea-fetch
+# @foxglove/just-fetch
 
 Isomorphic ponyfill wrapping native browser fetch and node-fetch.
 
 ### Usage
 
 ```ts
-import fetch from "@foxglove/panacea-fetch";
+import fetch from "@foxglove/just-fetch";
 ```
 
 Usage is the same whether you are targeting browsers or node.js. The browser
@@ -14,4 +14,4 @@ target returns the native `window.fetch` implementation, and in node.js the
 
 ### License
 
-@foxglove/panacea-fetch is licensed under [Mozilla Public License, v2.0](https://opensource.org/licenses/MPL-2.0).
+@foxglove/just-fetch is licensed under [Mozilla Public License, v2.0](https://opensource.org/licenses/MPL-2.0).
