+++ packages/just-fetch/browser/tsconfig.json
@@ -1,8 +1,8 @@
 {
-  "extends": "../../tsconfig.json",
-  "include": ["browser.ts"],
+  "extends": "../../../tsconfig.json",
+  "include": ["index.ts"],
   "compilerOptions": {
-    "outDir": "dist.browser",
+    "outDir": "../dist/browser",
     "target": "es2020",
     "module": "commonjs",
     "lib": ["es2020", "dom"],
