+++ packages/xmlrpc/src/XmlRpcClient.ts
@@ -2,7 +2,7 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-import fetch from "@foxglove/panacea-fetch";
+import fetch from "@foxglove/just-fetch";
 
 import { Deserializer } from "./Deserializer";
 import { serializeMethodCall } from "./Serializer";
